requirements.txt
networkx
setuptools
graspologic
wheel
aiofiles
aiohttp
configparser
future
nest_asyncio
numpy
pipmaster
pydantic
pydantic_settings
python-dotenv
numpy~=1.26.4
tenacity

# LLM packages
tiktoken
tqdm
xxhash
logfire
openai
chromadb
watchdog
pyvis
anthropic
typer
python-multipart
